+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `reject_nids` BOS option, to reject Sessions and Session Template which appear to reference NIDs.
+
 ### Changed
 - Sanitize BOS data during migration to this BOS version, to ensure it complies with the API specification.
   - For components and sessions, only validate the fields used to identify them (name, id, tenant). Delete them